backend/requirements.txt
asyncpg==0.28.0
SQLAlchemy==2.0.18

fastapi[all]==0.100.0
alembic==1.11.1
databases[postgresql]==0.4.2
fastapi-users[sqlalchemy]==12.0.0
fastapi-cache2[redis]==0.2.1

pytils==0.4.1
python-dotenv==1.0.0

celery==5.3.1
flower==2.0.0
redis==4.6.0

pytest==7.4.0
pytest-asyncio==0.21.0
